move createTravellersText
